Task 8 - Docker/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
joblib
pandas